+++ src/adlers/nlp/__init__.py
@@ -1,0 +1,1 @@
+"""Attention mechanisms for natural language processing."""
