+++ src/adlers/shared/_attention_base.py
