+++ src/adlers/time_series/__init__.py
@@ -1,0 +1,1 @@
+"""Attention mechanisms for time-series modeling."""
