+++ src/adlers/vision/__init__.py
@@ -1,0 +1,1 @@
+"""Attention mechanisms for computer vision."""
